accounts/abi: Prevent recalculation of internal fields (#20895)
This commit moves the complicated logic of how we assign method names
and event names if they already exist into their own functions for
better readability.

In this commit, I changed the way we calculate the string
representations, sig representations and the id's of methods. Before
that these fields would be recalculated everytime someone called .Sig()
.String() or .ID() on a method or an event.

Additionally this commit fixes issue #20856 as we assign names to inputs
with no name (input with name "" becomes "arg0")