Remove ExecutionResult
There is no easy way to classify the error returned from TransitionDb in Klaytn
is vmError or consensusError or just client side error(e.g. user specify gas limit which is lower than intrinsicGas).

Ethereum uses ExecutionResult because they can specify that errors in their codebase, but this is not adaptable to current Klaytn codebase.

The idea which classifies errors looks good, maybe we can apply it later.